==> 1_run_pipeline.ipynb <==
from classification_pipeline import classification_training_pipeline
from regression_pipeline import regression_training_pipeline


# set parameters
metric = 'HAMMING LOSS example based'
for learning_task in ['regression', 'pairwise_regression']:
    for task_output in ['single', 'multi']:
        regression_training_pipeline(metric, learning_task, task_output)


# classification 
classification_training_pipeline(metric, 'classification', 'single')
# pairwise_classification single output
classification_training_pipeline(metric, 'pairwise_classification', 'single')
# pairwise_classification multi output
classification_training_pipeline(metric, 'pairwise_classification', 'multi')
# cost sensitive pairwise classification singe target
classification_training_pipeline(metric, 'pairwise_classification', 'single', cost_sensitive=True)
