==> 4_heatmaps.ipynb <==
from utils import calculate_gap_closed
import pandas as pd
import numpy as np
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots



def gap_closed_heatmap(metrics):
    approaches = ['AS-R-MO', 'AS-R-SO', 'AS-PR-MO', 'AS-PR-SO', 'AS-C-SO', 'AS-PC-MO', 'AS-PC-SO', 'AS-CS-PC-SO']
    data = []
    for metric in metrics:
        data.append(calculate_gap_closed(metric, "results", 42))
    df = pd.DataFrame(data, index=metrics, columns=approaches)
    print(df)
    
    # Calculate the mean of each column and append it as a new row
    df.loc['Mean'] = df.mean()
    
    #rename metrics 
    index_mapping = {
        'HAMMING LOSS example based': 'Hamming loss example-based',
        'MACRO F1': 'F1 macro',
        'MICRO F1': 'F1 micro', 
        'AUCROC MICRO': 'AUROC micro', 
        'F1 example based': 'F1 example-based'
    }
    df.rename(index=index_mapping, inplace=True)

    # Plotting
    fig = px.imshow(df.round(2), color_continuous_scale='pubu', text_auto=True)
    fig.update_coloraxes(showscale=False)
    fig.update_xaxes(side="top")
    
    # Adjust layout
    fig.update_layout(
        autosize=True,
        width=720,
        margin=dict(
            l=0,
            r=0,
            b=0,
            t=0,
        ),
        font=dict(
            size=14
        )
    )
    
    # Drawing a dashed red line to separate the last row
    num_metrics = len(metrics)
    fig.add_shape(type="line",
                  x0=-0.5, y0=num_metrics - 0.5, x1=len(approaches) - 0.5, y1=num_metrics - 0.5,
                  line=dict(color="Red", width=4, dash="dash"),
                  layer="above"
                 )
    
    fig.show()
    fig.write_image(f'./figures/heatmap.pdf')
    # return df

gap_closed_heatmap(metrics=['HAMMING LOSS example based', 'MACRO F1', 'MICRO F1', 'AUCROC MICRO', 'F1 example based'])
